# smartreferencedata-web/src/main/webapp/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "smartreferencedata-api",
    "description": "SmartReferenceData REST API Information",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/smartreferencedata-api/refdatas/{id}": {
      "get": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Retrieve refdata by id",
        "operationId": "retrieveById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "You are not authorized to access the resource.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Update refdatas",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "You are not authorized to access the resource.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Delete refdatas by id",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/smartreferencedata-api/refdatas": {
      "get": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Retrieve refdatas",
        "operationId": "retrieveAll",
        "responses": {
          "200": {
            "description": "Successfully retrieved.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "You are not authorized to access the resource.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Register refdata",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "You are not authorized to access the resource.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/smartreferencedata-api/refdatas/type/{type}": {
      "get": {
        "tags": [
          "smart-reference-data-controller"
        ],
        "summary": "Retrieve refdatas by type",
        "operationId": "retrieveByRefDataType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "You are not authorized to access the resource.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceData"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The resource you were trying to reach is not found.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ReferenceData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "refDataCode": {
            "type": "string"
          },
          "refDataType": {
            "type": "string"
          },
          "refDataDescription": {
            "type": "string"
          },
          "refDataDescriptionDetail": {
            "type": "string"
          }
        },
        "description": "JSON with ReferenceDataDto object in and out"
      }
    }
  }
}